+++ apps/extension/src/utils/config/__tests__/example/v025.ts
@@ -1,5 +1,4 @@
 import type { TestSeriesObject } from './types'
-import type { Config } from '@/types/config/config'
 
 export const testSeries: TestSeriesObject = {
   'complex-config-from-v020': {
@@ -156,10 +155,10 @@ Translate to {{targetLang}}:
       betaExperience: {
         enabled: false,
       },
-    } satisfies Config,
+    },
   },
-  'config-from-v022': {
-    description: 'Add batch queue configuration for translation requests',
+  'config-with-no-default-openai-model': {
+    description: 'With no default openai model configured',
     config: {
       floatingButton: {
         disabledFloatingButtonPatterns: [],
@@ -184,44 +183,6 @@ Translate to {{targetLang}}:
           enabled: true,
           name: 'Microsoft Translator',
           provider: 'microsoft',
-        },
-        {
-          id: 'openai-default',
-          enabled: true,
-          apiKey: '123aqe',
-          baseURL: 'https://123',
-          models: {
-            read: {
-              customModel: null,
-              isCustomModel: false,
-              model: 'gpt-4.1-mini',
-            },
-            translate: {
-              customModel: 'gpt-4.1-m',
-              isCustomModel: true,
-              model: 'gpt-4.1-mini',
-            },
-          },
-          name: 'OpenAI',
-          provider: 'openai',
-        },
-        {
-          id: 'deepseek-default',
-          enabled: true,
-          models: {
-            read: {
-              customModel: null,
-              isCustomModel: false,
-              model: 'deepseek-chat',
-            },
-            translate: {
-              customModel: null,
-              isCustomModel: false,
-              model: 'deepseek-chat',
-            },
-          },
-          name: 'DeepSeek',
-          provider: 'deepseek',
         },
         {
           id: 'gemini-default',
@@ -251,7 +212,7 @@ Translate to {{targetLang}}:
         },
       ],
       read: {
-        providerId: 'deepseek-default',
+        providerId: 'gemini-default',
       },
       selectionToolbar: {
         enabled: true,
@@ -300,6 +261,6 @@ Translate to {{targetLang}}:
       betaExperience: {
         enabled: false,
       },
-    } satisfies Config,
+    },
   },
 }