+++ apps/website/src/app/[locale]/(home)/blog/[slug]/twitter-image.tsx
@@ -1,0 +1,1 @@
+export { alt, default, size } from './opengraph-image'